requirements.txt
chardet
cssmin
dropbox
google-api-python-client
gunicorn
httplib2
jinja2
markdown2
markupsafe
mysql-python
pygeoip
python-dateutil
python-memcached
pytz==2013d
requests
six
slugify
termcolor
werkzeug
semantic_version
gitpython==0.3.2.RC1